requirements.txt
black==21.12b0
Django==4.0.1
flake8==4.0.1
phonenumbers==8.12.41
Pillow==9.0.0
psycopg2-binary==2.9.3
python-decouple==3.5
